test: get storage when mount
